requirements.txt
flask>=3.0,<4.0
llama-cpp-python>=0.2.0
requests>=2.31
pyyaml>=6.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
psutil>=5.9
pySMART>=1.3
clamd>=1.0.2
python-registry>=1.4
sqlite-utils>=3.35
jinja2>=3.1
